+++ CHANGE_LOG.md
@@ -1,3 +1,22 @@
+# [1.35.0](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.34.0...v1.35.0) (2025-04-30)
+
+
+### Features
+
+* banner implementaion ([#647](https://github.com/oaknational/oak-ai-lesson-assistant/issues/647)) ([441c311](https://github.com/oaknational/oak-ai-lesson-assistant/commit/441c311a5586e1f75414830e1eea3b636e566f9b))
+
+# [1.34.0](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.33.0...v1.34.0) (2025-04-23)
+
+
+### Bug Fixes
+
+* don't throw on slack notification failure ([#621](https://github.com/oaknational/oak-ai-lesson-assistant/issues/621)) ([663989b](https://github.com/oaknational/oak-ai-lesson-assistant/commit/663989b734ec6ea498b920b55081096db50f029c))
+
+
+### Features
+
+* additional resource from partial aila lesson ([#644](https://github.com/oaknational/oak-ai-lesson-assistant/issues/644)) ([e7c1796](https://github.com/oaknational/oak-ai-lesson-assistant/commit/e7c179602c1d23a82e4c6a897caf82e05bfff27f))
+
 # [1.33.0](https://github.com/oaknational/oak-ai-lesson-assistant/compare/v1.32.0...v1.33.0) (2025-04-10)